ext2fs: Use correct type for sizes.
* ext2fs/ext2fs.h (ext2_{list,get,set}_xattr): Use 'size_t'.
* ext2fs/xattr.c (xattr_entry_{list,get,create): Use 'size_t' where
appropriate.
(ext2_{list,get,set}_xattr): Likewise.

+++ ext2fs/ext2fs.h
@@ -582,9 +582,9 @@ extern void ext2_warning (const char *, ...)
 /* ---------------------------------------------------------------- */
 /* xattr.c */
 
-error_t ext2_list_xattr (struct node *np, char *buffer, int *len);
-error_t ext2_get_xattr (struct node *np, const char *name, char *value, int *len);
-error_t ext2_set_xattr (struct node *np, const char *name, const char *value, int len, int flags);
+error_t ext2_list_xattr (struct node *np, char *buffer, size_t *len);
+error_t ext2_get_xattr (struct node *np, const char *name, char *value, size_t *len);
+error_t ext2_set_xattr (struct node *np, const char *name, const char *value, size_t len, int flags);
 error_t ext2_free_xattr_block (struct node *np);
 
 #endif
